requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
django-google-oauth==1.0.4
django-otp==1.6.1
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.8.0
google==3.0.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
multidict==6.7.0
oauth2-google-login==1.0.0
oauthlib==3.3.1
outcome==1.3.0.post0
phonenumbers==9.0.16
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9.1
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
trio==0.31.0
trio-websocket==0.12.2
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
yarl==1.22.0
Django>=4.0
gunicorn
psycopg2-binary
dj-database-url
python-decouple
whitenoise